🔥 Make movie separate component
